Support .pod files
This requires a couple of changes. First is to copy .pod files. The
other is to separate the name of the app from the name of the pod.
Podman apparently _always_ appends "-pod" to the name of the service,
despite the PodName parameter, which we might expect to control the name
of the service in addition to the name of the Podman object.