Introduce better way to update Tailscale
We don't usually care what version of Tailscale we have, so long as it
doesn't automatically change on us. But we also generally want _all_ the
services to use the same version, so it's not great when each repo has
the hard-coded version duplicated, especially when not all repos are
consuming the same version of podman-setup.

Instead, we'll have a separate command that will always fetch the latest
version (or a specified version, if needed). Then, we can just refer to
the "current" version and let Tailscale updates happen out of band from
the rest of the repo.