Tag image with app name
Using "current" was OK if each app had its own set of images, but when
multiple use the same image (such as Teslamate and Atuin both using
Postgres), they might need to ise there own image versions. With this
change, updating an image for one won't affect other consumers of the
image.

Many applications use Tailscale, and we still want to fetch one new
image to update all of them at once, so we'll explicitly apply the
"current" label to it.