Include .socket files
This is supposed to enable socket-activated services in systemd. It
doesn't exactly work, though, because if the service is
socket-activated, then it probably doesn't support being started all by
itself. And since it won't start by itself, the installation doesn't
actually reach the point of enabling the socket because starting the
service directly has already failed. So this needs a bit of work.